Transition to clap v3
Fix along the way `--mode` on write not working.

Change-type: patch
