+++ Cargo.toml
@@ -21,8 +21,8 @@ bitflags = "1"
 byteorder = "1"
 
 [dependencies.clap]
-version = "=2.33.0"
-default-features = false
+version = "3"
+features = ["derive"]
 
 [profile.release]
 opt-level = "z"
